Port menu hiding to fully React-state controlled
