+++ assets/scripts/info_bubble/info_bubble.js
@@ -37,7 +37,7 @@ import {
   undecorateWidth,
   processWidthInput
 } from '../util/width_units'
-import { isAnyMenuVisible, hideAllMenus } from '../menus/menu_old'
+import { isAnyMenuVisible, hideAllMenus } from '../menus/menu_controller'
 import { registerKeypress } from '../app/keypress'
 import { loseAnyFocus } from '../app/focus'
 import {
