+++ assets/scripts/main.js
@@ -31,7 +31,6 @@ import './gallery/scroll'
 import './gallery/view'
 import './info_bubble/info_bubble'
 import MenusContainer from './menus/MenusContainer'
-import './menus/menu_old'
 import './streets/name'
 import './streets/scroll'
 import './util/fetch_nonblocking'
