+++ assets/scripts/users/localization.js
@@ -4,7 +4,7 @@ import { ERRORS, showError } from '../app/errors'
 import { trackEvent } from '../app/event_tracking'
 import { checkIfEverythingIsLoaded } from '../app/initialization'
 import { MODES, getMode } from '../app/mode'
-import { hideAllMenus } from '../menus/menu_old'
+import { hideAllMenus } from '../menus/menu_controller'
 import { app } from '../preinit/app_settings'
 import { debug } from '../preinit/debug_settings'
 import { system } from '../preinit/system_capabilities'
